[#2/endpoints] Handle errors on endpoints

+++ src/Controller/Api/UserController.php
@@ -15,6 +15,7 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 /**
  * @Route("/api")
@@ -84,17 +85,24 @@ class UserController extends AbstractController
      * @Route("/users", name="new_user", methods={"POST"}, options={"expose" = true})
      * @param Request $request
      * @param ClientRepository $clientRepository
+     * @param ValidatorInterface $validator
      * @return Response
      */
-    public function new(Request $request, ClientRepository $clientRepository): Response
+    public function new(Request $request, ClientRepository $clientRepository, ValidatorInterface $validator): Response
     {
         $user = $this->serializer->deserialize($request->getContent(), User::class, 'json');
-        // TODO get the current client
-        $client = $clientRepository->find(25);
-        $user->setClient($client);
-        $this->entityManager->persist($user);
-        $this->entityManager->flush();
-        return new JsonResponse(["success" => "L'utilisateur a bien été créé"], 200);
+        $user->setClient($this->getUser());
+        $errors = $validator->validate($user);
+        if (!count($errors) > 0) {
+            $this->entityManager->persist($user);
+            $this->entityManager->flush();
+            return new JsonResponse(["success" => "L'utilisateur a bien été créé"], 200);
+        }
+        $errorMessages = [];
+        foreach ($errors as $error) {
+            $errorMessages[] = $error->getMessage();
+        }
+        return new JsonResponse($errorMessages, 202);
     }
 
     /**
@@ -107,20 +115,23 @@ class UserController extends AbstractController
     public function edit(int $id, Request $request, FormErrorsHandler $errorsHandler): Response
     {
         $user = $this->UserRepository->find($id);
-        $data = $this->serializer->deserialize($request->getContent(), "array", 'json');
-        $form = $this->createForm(UserType::class, $user, [
-            'csrf_protection' => false,
-            'method' => "PATCH"
-        ]);
-        $form->submit($data);
-        if ($form->isValid()) {
-            $this->entityManager->persist($user);
-            $this->entityManager->flush();
-            return new JsonResponse(["success" => "L'utilisateur a bien été modifié"], 200);
-        }
+        if ($user) {
+            $data = $this->serializer->deserialize($request->getContent(), "array", 'json');
+            $form = $this->createForm(UserType::class, $user, [
+                'csrf_protection' => false,
+                'method' => "PATCH"
+            ]);
+            $form->submit($data);
+            if ($form->isValid()) {
+                $this->entityManager->persist($user);
+                $this->entityManager->flush();
+                return new JsonResponse(["success" => "L'utilisateur a bien été modifié"], 200);
+            }
 
-        $errors = $errorsHandler->getErrors($form);
-        return new JsonResponse($errors, 400);
+            $errors = $errorsHandler->getErrors($form);
+            return new JsonResponse($errors, 202);
+        }
+        return new JsonResponse(["error" => "L'utilisateur n'existe pas"], 404);
     }
 
 
@@ -133,8 +144,11 @@ class UserController extends AbstractController
     public function delete(int $id, Request $request): Response
     {
         $user = $this->UserRepository->find($id);
-        $this->entityManager->remove($user);
-        $this->entityManager->flush();
-        return new JsonResponse(["success" => "L'utilisateur a bien été supprimé"], 200);
+        if ($user) {
+            $this->entityManager->remove($user);
+            $this->entityManager->flush();
+            return new JsonResponse(["success" => "L'utilisateur a bien été supprimé"], 200);
+        }
+        return new JsonResponse(["error" => "L'utilisateur n'existe pas"], 404);
     }
 }
